Some new features, some changes, added new branch
- add the capability of writing news and sending them as newsletters
-
move most user rank rights management into the Auth class
- make post
request more secure with link authentication
- enable admins to edit
preference texts with a wysiwg editor
- organize the preferences page
-
enable the deletion of user comments and users
- allow users to answer
to rumor and quote items
- make real 403 and 404 pages (with the right
headers set)
- manage pages with a PageManager class
- build an advanced
debug view for admins
- improve the javascript code
- slightly improve
the design
- improve the ratable user content model
- make the actions
sidebar autoupdatable
- create new actions page, useful for users of
phones and tablets who can't see the actions sidebar

These changes
aren't all - I'm sure that I forgot a bunch of them, as I wrote this
changes over a long period of time (longer than month). To prevent this
from hapening again, I created the new branch "develop" which I will be
updating quite often. The code of this branch wont be stable as it is
the current state of my code - the stable code is and will be in the
"master" branch.