Revised the user comment review process
Removed the ReviewText class with it's insult filter, as it doesn't work properly. Now, user comments aren't moderated at all except when comments of users which had been marked by the user he or she wants to commit or by an moderator or administrator.
Hopefully, it will fix #14, presenting the user an info text if it's comment has to moderated.